requirements.txt
selenium==3.14
pyyaml==3.11
requests==2.20